+++ CHANGELOG.md
@@ -525,6 +525,8 @@ Miscellaneous:
 - Connector table: a search filter is now available for the table that allows
   filtering across all columns.
 
+- Connectivity widget: larger input and partner sets can now be displayed.
+
 - Connector selection: a search filter is now available for the table that allows
   filtering across all columns.